requirements.txt
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dateutil>=2.8.0
schedule>=1.2.0
streamlit>=1.30.0
pandas>=2.0.0
fredapi>=0.5.0
yfinance>=0.2.0
plotly>=5.0.0
